CO2 Injection Modeling

Dolphyn.co2_injection — Method

co2_injection(EP::Model, inputs::Dict,setup::Dict)

This module defines the CO2 injection decision variable $x_{s,z,t}^{\textrm{C,INJ}} \forall k \in \mathcal{S}, z \in \mathcal{Z}, t \in \mathcal{T}$, representing CO2 injected into storage resource $s$ in zone $z$ at time period $t$.

The variable defined in this file named after vDAC\textunderscore{CO2}\textunderscore{Injected} covers all variables $x_{s,z,t}^{\textrm{C,INJ}}$.

This module defines the power consumption decision variable $x_{z,t}^{\textrm{E,INJ}} \forall z\in \mathcal{Z}, t \in \mathcal{T}$, representing power consumed by CO2 injection in zone $z$ at time period $t$.

The variable defined in this file named after vPower\textunderscore{CO2}\textunderscore{Injection} cover variable $x_{z,t}^{E,INJ}$.

Cost expressions

This module additionally defines contributions to the objective function from variable costs of CO2 injection (variable OM) from all resources over all time periods.

\[\begin{equation*} \textrm{C}^{\textrm{C,INJ,o}} = \sum_{s \in \mathcal{S}} \sum_{t \in \mathcal{T}} \omega_t \times \textrm{c}_{s}^{\textrm{INJ,VOM}} \times x_{s,z,t}^{\textrm{C,INJ}} \end{equation*}\]

Minimum and maximum injection output hourly

For resources where upper bound $\overline{x_{s}^{\textrm{C,INJ}}}$ of injection rate is defined, then we impose constraints on minimum and maximum injection rate

\[\begin{equation*} x_{s,z,t}^{\textrm{C,INJ}} \geq \underline{R_{s,z}^{\textrm{C,INJ}}} \times \overline{x_{s,z,t}^{\textrm{INJ}}} \quad \forall k \in \mathcal{S}, z \in \mathcal{Z}, t \in \mathcal{T} \end{equation*}\]

\[\begin{equation*} x_{s,z,t}^{\textrm{C,INJ}} \leq \overline{R_{s,z}^{\textrm{C,INJ}}} \times \overline{x_{s,z,t}^{\textrm{INJ}}} \quad \forall k \in \mathcal{S}, z \in \mathcal{Z}, t \in \mathcal{T} \end{equation*}\]

Maximum injection per year according to CO2 storage capacity per year

\[\begin{equation*} \sum_{t \in \mathcal{T}} x_{s,z,t}^{\textrm{C,INJ}} \leq y_{s,z}^{\textrm{C,STO}} \end{equation*}\]

CO2 Storage Investment

Dolphyn.co2_storage_investment — Method

co2_storage_investment(EP::Model, inputs::Dict, setup::Dict)

This module defines the total fixed cost (Investment + Fixed O&M) of the CO2 storage infrastructure in the CO2 supply chain

Sets up constraints common to all CO2 storage resources.

This function defines the expressions and constraints keeping track of total available CO2 storage capacity $y_{s}^{\textrm{C,STO}}$ (per year) as well as constraints on capacity.

The expression defined in this file named after vCapacity\textunderscore{CO2}\textunderscore{Storage}\textunderscore{per}\textunderscore{type} covers all variables $y_{s}^{\textrm{C,STO}}$.

The total capacity (per year) of each CO2 storage resource is defined as the sum of newly invested capacity (per year) based on the assumption there are no existing CO2 storage resources.

Cost expressions

This module additionally defines contributions to the objective function from investment costs of CO2 storage (fixed O\&M plus investment costs) from all resources $s \in \mathcal{S}$:

\[\begin{equation*} \textrm{C}^{\textrm{C,STO,c}} = \sum_{s \in \mathcal{S}} \sum_{z \in \mathcal{Z}} y_{s, z}^{\textrm{C,STO}}\times \textrm{c}_{s}^{\textrm{STO,INV}} + \sum_{s \in \mathcal{S}} \sum_{z \in \mathcal{Z}} y_{g, z}^{\textrm{C,STO,total}} \times \textrm{c}_{s}^{\textrm{STO,FOM}} \end{equation*}\]

Constraints on CO2 storage capacity (per year)

For resources where upper bound $\overline{y_{s}^{\textrm{C,STO}}}$ and lower bound $\underline{y_{s}^{\textrm{C,STO}}}$ of capacity is defined, then we impose constraints on minimum and maximum storage capacity (per year).

\[\begin{equation*} \underline{y_{s,z}^{\textrm{C,STO}}} \leq y_{s,z}^{\textrm{C,STO}} \leq \overline{y_{s,z}^{\textrm{C,STO}}} \quad \forall k \in \mathcal{S}, z \in \mathcal{Z} \end{equation*}\]