emission_hsc

DOLPHYN.emissions_hsc — Method

emissions_hsc(EP::Model, inputs::Dict, setup::Dict)

This function creates expression to add the CO2 emissions in each zone, which is subsequently added to the total emissions

h2_investment

DOLPHYN.h2_investment — Method

h2_discharge(EP::Model, inputs::Dict, UCommit::Int, Reserves::Int)

This module defines the production decision variable representing hydrogen injected into the network by resource $y$ by at time period $t$.

This module additionally defines contributions to the objective function from variable costs of generation (variable O&M plus fuel cost) from all resources over all time periods.

Variables

\[\begin{aligned} \vartheta _{k}^{GenNewCap}\ge 0 \end{aligned}\]

\[\begin{aligned} \vartheta _{k}^{GenRetCap}\ge 0 \end{aligned}\]

\[\begin{aligned} & \Delta^{total}_{y,z} =(\overline{\Delta_{y,z}}-\Delta_{y,z}) \forall y \in \mathcal{G}, z \in \mathcal{Z} \end{aligned}\]

One cannot retire more capacity than existing capacity.

\[\begin{aligned} &\Delta_{y,z} \leq \overline{\Delta_{y,z}} \hspace{4 cm} \forall y \in \mathcal{G}, z \in \mathcal{Z} \end{aligned}\]

In addition, this function adds investment and fixed O\&M related costs related to discharge/generation capacity to the objective function:

\[\begin{aligned} & \sum_{y \in \mathcal{G} } \sum_{z \in \mathcal{Z}} \left( (\pi^{INVEST}_{y,z} \times \overline{\Omega}^{size}_{y,z} \times ) + (\pi^{FOM}_{y,z} \times \Delta^{total}_{y,z})\right) \end{aligned}\]

h2nonserved_energy

DOLPHYN.h2_non_served_energy — Method

h2_non_served_energy(EP::Model, inputs::Dict, setup::Dict)

This function defines the non-served energy/curtailed demand decision variable $\Lambda_{s,t,z} \forall s \in \mathcal{S}, \forall t \in \mathcal{T}, z \in \mathcal{Z}$, representing the total amount of demand curtailed in demand segment $s$ at time period $t$ in zone $z$. The first segment of non-served energy, $s=1$, is used to denote the cost of involuntary demand curtailment (e.g. emergency load shedding or rolling blackouts), specified as the value of $n_{1}^{slope}$. Additional segments, $s \geq 2$ can be used to specify a segment-wise approximation of a price elastic demand curve, or segments of price-responsive curtailable loads (aka demand response). Each segment denotes a price/cost at which the segment of demand is willing to curtail consumption, $n_{s}^{slope}$, representing the marginal willingness to pay for electricity of this segment of demand (or opportunity cost incurred when demand is not served) and a maximum quantity of demand in this segment, $n_{s}^{size}$, specified as a share of demand in each zone in each time step, $D_{t,z}.$ Note that the current implementation assumes demand segments are an equal share of hourly load in all zones.

This function defines contributions to the objective function from the cost of non-served energy/curtailed demand from all demand curtailment segments $s \in \mathcal{S}$ over all time periods $t \in \mathcal{T}$ and all zones $z \in \mathcal{Z}$:

\[\begin{aligned} Obj_{NSE} = \sum_{s \in \mathcal{S} } \sum_{t \in \mathcal{T}} \sum_{z \in \mathcal{Z}}\omega_{t} \times n_{s}^{slope} \times \Lambda_{s,t,z} \end{aligned}\]

Contributions to the power balance expression from non-served energy/curtailed demand from each demand segment $s \in \mathcal{S}$ are also defined as:

\[\begin{aligned} PowerBal_{NSE} = \sum_{s \in \mathcal{S} } \Lambda_{s,t,z} \hspace{4 cm} \forall s \in \mathcal{S}, t \in \mathcal{T} \end{aligned}\]

Bounds on curtailable demand

Demand curtailed in each segment of curtailable demands $s \in \mathcal{S}$ cannot exceed maximum allowable share of demand:

\[\begin{aligned} \Lambda_{s,t,z} \leq (n_{s}^{size} \times D_{t,z}) \hspace{4 cm} \forall s \in \mathcal{S}, t \in \mathcal{T}, z\in \mathcal{Z} \end{aligned}\]

Additionally, total demand curtailed in each time step cannot exceed total demand:

\[\begin{aligned} \sum_{s \in \mathcal{S} } \Lambda_{s,t,z} \leq D_{t,z} \hspace{4 cm} \forall t \in \mathcal{T}, z\in \mathcal{Z} \end{aligned}\]

h2_outputs

Missing docstring.

Missing docstring for DOLPHYN.h2_outputs(EP::Model, inputs::Dict, setup::Dict). Check Documenter's build log for details.