Hydrogen Gas to Power

Hydrogen Gas to Power All

DOLPHYN.h2_g2p_all — Method

h2_g2p_all(EP::Model, inputs::Dict, setup::Dict)

The hydrogen to power module creates decision variables, expressions, and constraints related to hydrogen generation infrastructure

Investment and FOM cost expression, VOM cost expression, minimum and maximum capacity limits

Hydrogen Gas to Power Commitment

DOLPHYN.h2_g2p_commit — Method

h2_g2p_commit(EP::Model, inputs::Dict, setup::Dict)

This module creates decision variables, expressions, and constraints related to various hydrogen to power technologies with unit commitment constraints (e.g. natural gas reforming etc.)

Documentation to follow ******

Hydrogen Gas to Power Discharge

DOLPHYN.h2_g2p_discharge — Method

h2_g2p_discharge(EP::Model, inputs::Dict, setup::Dict)

This module defines the production decision variable representing power form hydrogen injected into the network by resource $y$ by at time period $t$.

This module additionally defines contributions to the objective function from variable costs of generation (variable O&M plus fuel cost) from all resources over all time periods.

Hydrogen Gas to Power Investment

DOLPHYN.h2_g2p_investment — Method

h2_g2p_investment(EP::Model, inputs::Dict, setup::Dict)

This module defines the investment cost of hydrogen to power plants.

Hydrogen Gas to Power No Commitment

DOLPHYN.h2_g2p_no_commit — Method

h2_g2p_no_commit(EP::Model, inputs::Dict,setup::Dict)

This module creates decision variables, expressions, and constraints related to various hydrogen to power technologies without unit commitment constraints

Hydrogen Gas to Power

DOLPHYN.h2_g2p — Method

h2_g2p(EP::Model, inputs::Dict, setup::Dict)

This module creates decision variables, expressions, and constraints related to various hydrogen to power technologies.