Hydrogen Storage

DOLPHYN.h2_storage — Method

h2_storage(EP::Model, inputs::Dict, setup::Dict)

A wide range of energy storage devices (all $s \in \mathcal{S}$) can be modeled in DOLPHYN, using one of two generic storage formulations: (1) storage technologies with symmetric charge and discharge capacity (all $s \in \mathcal{S}^{sym}$), such as Lithium-ion batteries and most other electrochemical storage devices that use the same components for both charge and discharge; and (2) storage technologies that employ distinct and potentially asymmetric charge and discharge capacities (all $s \in \mathcal{S}^{asym}$), such as most thermal storage technologies or hydrogen electrolysis/storage/fuel cell or combustion turbine systems.