Hydrogen Transmission

DOLPHYN.h2_pipeline — Method

h2_pipeline(EP::Model, inputs::Dict, setup::Dict)

This function includes the variables, expressions and objective funtion of hydrogen pipeline.

This function expresses hydrogen exchange through pipeline i between two zones and can be split into H2 delivering and flowing out.

This module defines the hydrogen pipeline construction decision variable $y_{i,z \rightarrow z^{\prime}}^{\textrm{H,PIP}} \forall i \in \mathcal{I}, z \rightarrow z^{\prime} \in \mathcal{B}$, representing newly constructed hydrogen pipeline of type $i$ through path $z \rightarrow z^{\prime}$.

This module defines the hydrogen pipeline flow decision variable $x_{i,z \rightarrow z^{\prime},t}^{\textrm{H,PIP}} \forall i \in \mathcal{I}, z \rightarrow z^{\prime} \in \mathcal{B}, t \in \mathcal{T}$, representing hydrogen flow via pipeline of type $i$ through path $z \rightarrow z^{\prime}$ at time period $t$.

This module defines the hydrogen pipeline storage level decision variable $U_{i,z \rightarrow z^{\prime},t}^{\textrm{H,PIP}} \forall i \in \mathcal{I}, z \rightarrow z^{\prime} \in \mathcal{B}, t \in \mathcal{T}$, representing hydrogen stored in pipeline of type $i$ through path $z \rightarrow z^{\prime}$ at time period $t$.

The variable defined in this file named after vH2NPipe covers variable $y_{i,z \rightarrow z^{\prime}}^{\textrm{H,PIP}}$.

The variable defined in this file named after vH2PipeFlow_pos covers variable $x_{i,z \rightarrow z^{\prime},t}^{\textrm{H,PIP+}}$.

The variable defined in this file named after vH2PipeFlow_neg covers variable $x_{i,z \rightarrow z^{\prime},t}^{\textrm{H,PIP-}}$.

The variable defined in this file named after vH2PipeLevel covers variable $U_{i,z \rightarrow z^{\prime},t}^{\textrm{H,PIP}}$.

Cost expressions

This module additionally defines contributions to the objective function from investment costs of generation (fixed OM plus construction) from all pipeline resources $i \in \mathcal{I}$:

\[\begin{equation*} \textrm{C}^{\textrm{H,PIP,c}}=\delta_{i}^{\textrm{H,PIP}} \sum_{i \in \mathbb{I}} \sum_{z \rightarrow z^{\prime} \in \mathbb{B}} \textrm{c}_{i}^{\textrm{H,PIP}} \textrm{L}_{z \rightarrow z^{\prime}} l_{i,z \rightarrow z^{\prime}} h_{i,z \rightarrow z^{\prime}, t}^{\textrm{H,PIP}}=h_{i, z \rightarrow z^{\prime}, t}^{\textrm{H,PIP+}}-h_{i, z \rightarrow z^{\prime}, t}^{\textrm{PIP-}} \quad \forall i \in \mathbb{I}, z \rightarrow z^{\prime} \in \mathbb{B}, t \in \mathbb{T} \end{equation*} ``` The flow rate of H2 through pipeline type $i$ is capped by the operational limits of the pipeline, multiplied by the number of constructed pipeline $i$\]

math \begin{equation} \overline{\textrm{F}}{i} l{i,z \rightarrow z^{\prime}} \geq x{i,z \rightarrow z^{\prime}, t}^{\textrm{\textrm{H,PIP+}}}, x{i,z \rightarrow z^{\prime}, t}^{\textrm{\textrm{H,PIP-}}} \geq 0 \quad \forall i \in \mathbb{I}, z \rightarrow z^{\prime} \in \mathbb{B}, t \in \mathbb{T} \end{equation}

The pipeline has storage capacity via line packing:

math \begin{equation} \overline{\textrm{U}}{i}^{\textrm{\textrm{H,PIP}}} l{i,z \rightarrow z^{\prime}} \geq -\sum{\tau=t{0}}^{t}\left(x{i,z^{\prime} \rightarrow z, \tau}^{\textrm{\textrm{H,PIP}}}+x{i,z \rightarrow z^{\prime}, \tau}^{\textrm{\textrm{H,PIP}}}\right) \Delta t \geq \underline{\textrm{R}}{i}^{\textrm{\textrm{H,PIP}}} \overline{\textrm{E}}{i}^{\textrm{\textrm{H,PIP}}} l_{i,z \rightarrow z^{\prime}} \ & \forall z^{\prime} \in \mathbb{Z}, z \in \mathbb{Z}, i \in \mathbb{I}, t \in \mathbb{T} \end{equation}

The change of hydrogen pipeline storage inventory is modeled as follows:

math \begin{equation} U{i,z \rightarrow z^{\prime},t}^{\textrm{H,PIP}} - U{i,z \rightarrow z^{\prime},t-1} = x{i,z \rightarrow z^{\prime},t}^{\textrm{H,PIP-}} + x{i,z^{\prime} \rightarrow z,t}^{\textrm{H,PIP-}} \end{equation} ```