Hydrogen Truck All

DOLPHYN.h2_truck_all — Method

h2_truck_all(EP::Model, inputs::Dict, setup::Dict)

This function defines a series of operating variables,expressions and constraints in truck scheduling and routing model.

Variables

The sum of full and empty trucks should equal the total number of invested trucks.

\[\begin{equation*} v_{j, t}^{\textrm{F}}+v_{j, t}^{\textrm{E}}=V_{j} \quad \forall j \in \mathbb{J}, t \in \mathbb{T} \end{equation*} \]

The full (empty) trucks include full (empty) trucks in transit and staying at each zones.

\[\begin{aligned} v_{j, t}^{\textrm{F}}=\sum_{z \rightarrow z^{\prime} \in \mathbb{B}} u_{z \rightarrow z^{\prime}, t}^{\textrm{F}}+\sum_{z \in \mathbb{Z}} q_{z, j, t}^{\textrm{F}} \\ v_{j, t}^{\textrm{E}}=\sum_{z \rightarrow z^{\prime} \in \mathbb{B}} u_{z \rightarrow z^{\prime}, t}^{\textrm{E}}+\sum_{z \in \mathbb{Z}} q_{z, j, t}^{\textrm{E}} \quad \forall j \in \mathbb{J}, t \in \mathbb{T} \end{aligned} \]

Expressions

The change of the total number of full (empty) available trucks at zone z should equal the number of charged (discharged) trucks minus the number of discharged (charged) trucks at zone z plus the number of full (empty) trucks that just arrived minus the number of full (empty) trucks that just departed:

\[\begin{aligned} q_{z, j, t}^{\textrm{F}}-q_{z, j, t-1}^{\textrm{F}}=& q_{z, j, t}^{\textrm{CHA}}-q_{z, j, t}^{\textrm{DIS}} \\ &+\sum_{z^{\prime} \in \mathbb{Z}}\left(-x_{z \rightarrow z^{\prime}, j, t-1}^{\textrm{F}}+y_{z \rightarrow z^{\prime}, j, t-1}^{\textrm{F}}\right) \\ q_{z, j, t}^{\textrm{E}}-q_{z, j, t-1}^{\textrm{E}}=&-q_{z, j, t}^{\textrm{CHA}}+q_{z, j, t}^{\textrm{DIS}} \\ &+\sum_{z^{\prime} \in \mathbb{Z}}\left(-x_{z \rightarrow z^{\prime}, j, t-1}^{\textrm{E}}+y_{z \rightarrow z^{\prime} j, t-1}^{\textrm{E}}\right) \\ \quad \forall z \in \mathbb{Z}, j \in \mathbb{J}, t \in \mathbb{T} \end{aligned}\]

The change of the total number of full (empty) trucks in transit from zone z to zone zz should equal the number of full (empty) trucks that just departed from zone z minus the number of full (empty) trucks that just arrived at zone zz:

\[\begin{aligned} u_{z \rightarrow z^{\prime}, j, t}^{\textrm{F}}-u_{z \rightarrow z^{\prime}, j, t-1}^{\textrm{F}} & =x_{z \rightarrow z^{\prime}, j, t-1}^{\textrm{F}}-y_{z \rightarrow z^{\prime}, j, t-1}^{\textrm{F}} \\ u_{z \rightarrow z^{\prime}, j, t}^{\textrm{E}}-u_{z \rightarrow z^{\prime}, j, t-1}^{\textrm{E}} & =x_{z \rightarrow z^{\prime}, j, t-1}^{\textrm{E}}-y_{z \rightarrow z^{\prime}, j, t-1}^{\textrm{E}} \\ & \quad \forall z \rightarrow z^{\prime} \in \mathbb{B}, j \in \mathbb{J}, t \in \mathbb{T} \end{aligned} \]

The amount of H2 delivered to zone z should equal the truck capacity times the number of discharged trucks minus the number of charged trucks, adjusted by theH2 boil-off loss during truck transportation and compression.

\[\begin{aligned} h_{z, j, t}^{\textrm{H,TRU}}=\left[\left(1-\sigma_{j}\right) q_{z, j, t}^{\textrm{DIS}}-q_{z, j, t}^{\textrm{CHA}}\right] \overline{\textrm{E}}_{j}^{\textrm{H,TRU}} \\ \quad \forall z \rightarrow z^{\prime} \in \mathbb{B}, j \in \mathbb{J}, t \in \mathbb{T} \end{aligned} \]

The minimum travelling time delay is modelled as follows.

\[\begin{aligned} u_{z \rightarrow z^{\prime}, j, t}^{\textrm{F}} \geq \sum_{e=t-\Delta_{z \rightarrow z^{\prime}+1}}^{e=t} x_{z \rightarrow z^{\prime}, j, e}^{\textrm{F}} \\ u_{z \rightarrow z^{\prime}, j, t}^{\textrm{E}} \geq \sum_{e=t-\Delta_{z \rightarrow z^{\prime}+1}}^{e=t} x_{z \rightarrow z, j, e}^{\textrm{E}} \quad \forall z \rightarrow z^{\prime} \in \mathbb{B}, j \in \mathbb{J}, t \in \mathbb{T} \end{aligned}\]

\[\begin{aligned} u_{z \rightarrow z^{\prime}, j, t}^{\textrm{F}} \geq \sum_{e=t+1}^{e=t+\Delta_{z \rightarrow z^{\prime}}} y_{z \rightarrow z^{\prime} j, e}^{\textrm{F}} \\ u_{z \rightarrow z, j, t}^{\textrm{E}} \geq \sum_{e=t+1}^{e=t+\Delta_{z \rightarrow z^{\prime}}} y_{z \rightarrow z^{\prime} j, e}^{\textrm{E}} \\ \quad \forall z \rightarrow z^{\prime} \in \mathbb{B}, j \in \mathbb{J}, t \in \mathbb{T} \end{aligned} \]

Constraints

The charging capability of truck stations is limited by their compression or liquefaction capacity.

\[\begin{equation*} q_{z, j, t}^{\textrm{CHA}} \overline{\textrm{E}}_{j}^{\textrm{H,TRU}} \leq H_{z, j}^{\textrm{H,TRU}} \quad \forall z \in \mathbb{Z}, j \in \mathbb{J}, t \in \mathbb{T} \end{equation*} \]