Synthetic Fuels Resources Investment

Dolphyn.syn_fuel_investment — Method

syn_fuel_investment(EP::Model, inputs::Dict, setup::Dict)

Sets up constraints common to all synthetic fuels resources.

This function defines the expressions and constraints keeping track of total available synthetic fuels capacity $y_{f}^{\textrm{C,Syn}}$ based on its input CO2 in tonne per hour as well as constraints on capacity.

The expression defined in this file named after vCapacity\textunderscore{Syn}\textunderscore{Fuel}\textunderscore{per}\textunderscore{type} covers all variables $y_{f}^{\textrm{C,Syn}}$.

The total capacity of each synthetic fuels resource is defined as the sum of newly invested capacity based on the assumption there are no existing synthetic fuels resources.

Cost expressions

This module additionally defines contributions to the objective function from investment costs of synthetic fuels (fixed O\&M plus investment costs) from all generation resources $f \in \mathcal{F}$:

\[\begin{equation*} \textrm{C}^{\textrm{LF,Syn,c}} = \sum_{f \in \mathcal{F}} \sum_{z \in \mathcal{Z}} y_{f, z}^{\textrm{C,Syn}}\times \textrm{c}_{f}^{\textrm{Syn,INV}} + \sum_{f \in \mathcal{F}} \sum_{z \in \mathcal{Z}} y_{f, z}^{\textrm{C,Syn}} \times \textrm{c}_{f}^{\textrm{Syn,FOM}} \end{equation*}\]

Constraints on synthetic fuels resource capacity

For resources where upper bound $\overline{y_{f}^{\textrm{C,Syn}}}$ and lower bound $\underline{y_{f}^{\textrm{C,Syn}}}$ of capacity is defined, then we impose constraints on minimum and maximum synthetic fuels resource input CO2 capacity.

\[\begin{equation*} \underline{y_{f}^{\textrm{C,Syn}}} \leq y_{f}^{\textrm{C,Syn}} \leq \overline{y_{f}^{\textrm{C,Syn}}} \quad \forall f \in \mathcal{F} \end{equation*}\]

Synthetic Fuels Variable Cost

Dolphyn.syn_fuel_outputs — Method

syn_fuel_outputs(EP::Model, inputs::Dict, setup::Dict)

Sets up variables common to all synthetic fuels resources.

This module defines the synthetic fuels resource decision variable $x_{f,t}^{\textrm{C,Syn}} \forall f \in \mathcal{F}, z \in \mathcal{Z}, t \in \mathcal{T}$, representing CO2 input into the synthetic fuels resource $f$ at time period $t$.

\[x_{f,b,t}^{\textrm{By,Syn}} \forall f \in \mathcal{F}, \forall b \in \mathcal{B}, z \in \mathcal{Z}, t \in \mathcal{T}\]

, representing synthetic fuels by products $b$ (if any) by the synthetic fuels resource $f$ at time period $t$.

The variables defined in this file named after vSFCO2in covers all variables $x_{f,t}^{\textrm{C,Syn}}$ and vSFByProd covers all variables $x_{f,b,t}^{\textrm{Syn,By}}$.

Cost expressions

This module additionally defines contributions to the objective function from variable costs of generation (variable OM plus fuel cost) from all synthetic fuels resources over all time periods.

\[\begin{equation*} \textrm{C}^{\textrm{LF,Syn,o}} = \sum_{f \in \mathcal{F}} \sum_{t \in \mathcal{T}} \omega_t \times \left(\textrm{c}_{f}^{\textrm{Syn,VOM}} + \textrm{c}_{f}^{\textrm{Syn,FUEL}}\right) \times x_{f,t}^{\textrm{C,Syn}} \end{equation*}\]

This module also defines contributions to the objective function from revenues of by-products (if any) of synthetic fuels generation from all resources over all time periods.

\[\begin{equation*} \textrm{C}^{\textrm{LF,Syn,r}} = \sum_{f \in \mathcal{F}} \sum_{b \in \mathcal{B}} \sum_{t \in \mathcal{T}} \omega_t \times x_{f,b,t}^{\textrm{By,Syn}} \times \textrm{c}_{b}^{\textrm{By,Syn}} \end{equation*}\]

Emissions

Dolphyn.liquid_fuel_emissions — Method

liquid_fuel_emissions(EP::Model, inputs::Dict, setup::Dict)

This function creates expression to add the CO2 emissions for liquid fuels supply chain in each zone, which is subsequently added to the total emissions.

These include emissions from synthetic fuels production and by-products (if any), as well as combustion of each type of conventional and synthetic fuels (gasoline, jetfuel and diesel).