Load Hydrogen Inputs

DOLPHYN.load_h2_inputs — Method

load_h2_inputs(inputs::Dict,setup::Dict,path::AbstractString)

Loads various data inputs from multiple input .csv files in path directory and stores variables in a Dict (dictionary) object for use in model() function

inputs: inputs - dict object containing input data setup - dict object containing setup parameters path - string path to working directory

returns: Dict (dictionary) object containing all data inputs of hydrogen sector.

Load Carbon Cap in Hydrogen Sector

DOLPHYN.load_co2_cap_hsc — Method

load_co2_cap_hsc(setup::Dict, path::AbstractString, sep::AbstractString, inputs_co2_hsc::Dict)

Function for reading input parameters related to CO$_2$ emissions cap constraints in hydrogen supply chain only.

Load Hydrogen Demand

DOLPHYN.load_h2_demand — Method

load_h2_demand(setup::Dict, path::AbstractString, sep::AbstractString, inputs_load::Dict)

Function for reading input parameters related to hydrogen load (demand) of each zone.

Load Hydrogen Gas to Power Variability

DOLPHYN.load_h2_g2p_variability — Method

load_h2_g2p_variability(setup::Dict, path::AbstractString, sep::AbstractString, inputs_genvar::Dict)

Function for reading input parameters related to hourly maximum capacity factors for all hydrogen to power plants.

Load Hydrogen Gas to Power

DOLPHYN.load_h2_g2p — Method

load_h2_g2p(setup::Dict, path::AbstractString, sep::AbstractString, inputs_gen::Dict)

Function for reading input parameters related to hydrogen to power generators.

Load Hydrogen Generators

DOLPHYN.load_h2_gen — Method

load_h2_gen(setup::Dict, path::AbstractString, sep::AbstractString, inputs_gen::Dict)

Function for reading input parameters related to hydrogen generators.

Load Hydrogen Generators Variability

DOLPHYN.load_h2_generators_variability — Method

load_h2_generators_variability(setup::Dict, path::AbstractString, sep::AbstractString, inputs_genvar::Dict)

Function for reading input parameters related to hourly maximum capacity factors for all hydrogen generators (plus storage).

Load Hydrogen Pipelines

DOLPHYN.load_h2_pipeline_data — Method

load_h2_pipeline_data(setup::Dict, path::AbstractString, sep::AbstractString, inputs_nw::Dict)

Function for reading input parameters related to the hydrogen transmission network via pipelines.

Load Hydrogen Truck

DOLPHYN.load_h2_truck — Method

load_h2_truck(path::AbstractString, sep::AbstractString, inputs_truck::Dict)

Function for reading input parameters related to hydrogen trucks.