Reading Input Files

DOLPHYN.load_inputs — Method

load_inputs(setup::Dict,path::AbstractString)

Loads various data inputs from multiple input .csv files in path directory and stores variables in a Dict (dictionary) object for use in model() function

inputs: setup - dict object containing setup parameters path - string path to working directory

returns: Dict (dictionary) object containing all data inputs

Fuels Data

DOLPHYN.load_fuels_data — Method

load_fuels_data(setup::Dict, path::AbstractString, sep::AbstractString, inputs_fuel::Dict)

Function for reading input parameters related to fuel costs and CO$_2$ emission intensity of fuels.

Generators Input Data

DOLPHYN.load_generators_data — Method

load_generators_data(setup::Dict, path::AbstractString, sep::AbstractString, inputs_gen::Dict, fuel_costs::Dict, fuel_CO2::Dict)

Function for reading input parameters related to electricity generators (plus storage and flexible demand resources).

Variability of Generators' Outputs

DOLPHYN.load_generators_variability — Method

load_generators_variability(setup::Dict, path::AbstractString, sep::AbstractString, inputs_genvar::Dict)

Function for reading input parameters related to hourly maximum capacity factors for all generators (plus storage and flexible demand resources).

Load Data

DOLPHYN.load_load_data — Method

load_load_data(setup::Dict, path::AbstractString, sep::AbstractString, inputs_load::Dict)

Function for reading input parameters related to electricity load (demand).

Transmission Network

Minimum Capacity Requirements

DOLPHYN.load_minimum_capacity_requirement — Method

load_minimum_capacity_requirement(path::AbstractString,sep::AbstractString, inputs::Dict, setup::Dict)

Function for reading input parameters related to mimimum capacity requirement constraints (e.g. technology specific deployment mandates).

Capacity Reserve Margin

DOLPHYN.load_cap_reserve_margin — Method

load_cap_reserve_margin(setup::Dict, path::AbstractString, sep::AbstractString, inputs_crm::Dict)

Function for reading input parameters related to planning reserve margin constraints.

CO$_2$ Emissions Cap

DOLPHYN.load_co2_cap — Method

load_co2_cap(setup::Dict, path::AbstractString, sep::AbstractString, inputs_co2::Dict)

Function for reading input parameters related to CO$_2$ emissions cap constraints.

Energy Share Requirement

DOLPHYN.load_energy_share_requirement — Method

load_energy_share_requirement(setup::Dict, path::AbstractString, sep::AbstractString, inputs_ESR::Dict)

Function for reading input parameters related to mimimum energy share requirement constraints (e.g. renewable portfolio standard or clean electricity standard policies).

Mapping Representative Time Periods

DOLPHYN.load_period_map — Method

load_period_map(setup::Dict,path::AbstractString,sep::AbstractString, inputs::Dict)

Function for reading input parameters related to mapping of representative time periods to full chronological time series.

Reserve

DOLPHYN.load_reserves — Method

load_reserves(setup::Dict,path::AbstractString,sep::AbstractString, inputs_res::Dict)

Function for reading input parameters related to frequency regulation and operating reserve requirements.