Storage All

DOLPHYN.storage_all — Method

storage_all(EP::Model, inputs::Dict, Reserves::Int, OperationWrapping::Int)

Sets up variables and constraints common to all storage resources.

Storage discharge and inventory level decision variables

This module defines the storage energy inventory level variable $U_{s,z,t}^{\textrm{E,STO}}\quad \forall s \in \mathcal{S}, z \in \mathcal{Z}, t \in \mathcal{T}$, representing energy stored in the storage device $s$ in zone $z$ at time period $t$.

This module defines the power charge decision variable $x_{s,z,t}^{\textrm{E,CHA}}\quad \forall s \in \mathcal{S}, z \in \mathcal{Z}, t \in \mathcal{T}$, representing charged power into the storage device $s$ in zone $z$ at time period $t$.

The variable defined in this file named after vS covers $U_{s,z,t}^{\textrm{E,STO}}$.

The variable defined in this file named after vCHARGE covers $x_{s,z,t}^{\textrm{E,CHA}}$.

Cost expressions

This module additionally defines contributions to the objective function from variable costs (variable O\&M plus fuel cost) of charging action of storage devices $s \in \mathcal{S}$ over all time periods $t \in \mathcal{T}$:

\[\begin{equation*} \textrm{C}^{\textrm{E,STO,o}} = \sum_{s \in \mathcal{S}} \sum_{z \in \mathcal{Z}} \sum_{t \in \mathcal{T}}\omega_t \times \textrm{c}_{s,z,t}^{\textrm{E,STO,o}} \times x_{s,z,t}^{\textrm{E,CHA}} \end{equation*}\]

Power balance expressions

Contributions to the power balance expression from storage charging and discharging action from storage devices $s \in \mathcal{S}$ are also defined as:

\[\begin{equation*} PowerBal_{STO} = \sum_{s \in \mathcal{S}} \left(x_{s,z,t}^{\textrm{E,DIS}} - x_{s,z,t}^{\textrm{E,CHA}}\right) \quad \forall z \in \mathcal{Z}, t \in \mathcal{T} \end{equation*}\]

Storage inventory level track constraints

The following constraints apply to all storage resources, $s \in \mathcal{S}$, regardless of whether the charge/discharge capacities are symmetric or asymmetric.

The following two constraints track the state of charge of the storage resources at the end of each time period, relating the volume of energy stored at the end of the time period, $U_{s,z,t}^{\textrm{E,STO}}$, to the state of charge at the end of the prior time period, $U_{s,z,t-1}^{\textrm{E,STO}}$, the charge and discharge decisions in the current time period, $x_{s,z,t}^{\textrm{E,CHA}}, x_{s,z,t}^{\textrm{E,DIS}}$, and the self discharge rate for the storage resource (if any), $\eta_{s,z}^{\textrm{E,loss}}$. The first of these two constraints enforces storage inventory balance for interior time steps $(t \in \mathcal{T}^{interior})$, while the second enforces storage balance constraint for the initial time step $(t \in \mathcal{T}^{start})$.

\[\begin{aligned} U_{s,z,t}^{\textrm{E,STO}} &= U_{s,z,t-1}^{\textrm{E,STO}} - \frac{1}{\eta_{s,z}^{\textrm{E,STO}}}x_{s,z,t}^{\textrm{E,DIS}} + \eta_{s,z}^{\textrm{E,STO}}x_{s,z,t}^{\textrm{E,STO}} - \eta_{s,z}^{\textrm{E,loss}}U_{s,z,t-1} \quad \forall s \in \mathcal{S}, z \in \mathcal{Z}, t \in \mathcal{T}^{interior} \\ U_{s,z,t}^{\textrm{E,STO}} &= U_{s,z,t+\tau^{period}-1}^{\textrm{E,STO}} - \frac{1}{\eta_{s,z}^{\textrm{E,STO}}}x_{s,z,t}^{\textrm{E,DIS}} + \eta_{s,z}^{\textrm{E,STO}}x_{s,z,t}^{\textrm{E,CHA}} - \eta_{s,z}^{\textrm{E,loss}} U_{s,z,t+\tau^{period}-1} \quad \forall s \in \mathcal{S}, z \in \mathcal{Z}, t \in \mathcal{T}^{start} \end{aligned}\]

When modeling the entire year as a single chronological period with total number of time steps of $\tau^{period}$, storage inventory in the first time step is linked to storage inventory at the last time step of the period representing the year. Alternatively, when modeling the entire year with multiple representative periods, this constraint relates storage inventory in the first timestep of the representative period with the inventory at the last time step of the representative period, where each representative period is made of $\tau^{period}$ time steps. In this implementation, energy exchange between representative periods is not permitted. When modeling representative time periods, DOLPHYN enables modeling of long duration energy storage which tracks state of charge between representative periods enable energy to be moved throughout the year. If LongDurationStorage=1 and OperationWrapping=1, this function calls long_duration_storage() in long_duration_storage.jl to enable this feature.

Bounds on storage power and energy capacity

The storage power capacity sets lower and upper bounds on the storage energy capacity due to charging or discharging duration.

\[\begin{aligned} y_{s,z}^{\textrm{E,STO,POW}} \times \tau_{s,z}^{MinDuration} &\leq y_{s,z}^{\textrm{E,STO,ENE}} \\ y_{s,z}^{\textrm{E,STO,POW}} \times \tau_{s,z}^{MaxDuration} &\geq y_{s,z}^{\textrm{E,STO,ENE}} \end{aligned}\]

It limits the volume of energy $U_{s,z,t}^{\textrm{E,STO}}$ at any time $t$ to be less than the installed energy storage capacity $y_{s,z}^{\textrm{E,STO,ENE}}$.

\[\begin{equation*} 0 \leq U_{s,z,t}^{\textrm{E,STO}} \leq y_{s,z}^{\textrm{E,STO,ENE}} \quad \forall s \in \mathcal{S}, z \in \mathcal{Z}, t \in \mathcal{T} \end{equation*}\]

It also limits the discharge power $x_{s,z,t}^{\textrm{E,DIS}}$ at any time to be less than the installed power capacity $y_{s,z}^{\textrm{E,STO},POW}$. Finally, the maximum discharge rate for storage resources, $x_{s,z,t}^{\textrm{E,STO}}$, is constrained to be less than the discharge power capacity, $y_{s,z}^{\textrm{E,STO},POW}$ or the state of charge at the end of the last period, $U{s,z,t-1}^{\textrm{E,STO}}$, whichever is less.

\[\begin{aligned} 0 &\leq x_{s,z,t}^{\textrm{E,DIS}} \leq y_{s,z}^{\textrm{E,STO,POW}} \quad \forall s \in \mathcal{S}, z \in \mathcal{Z}, t \in \mathcal{T} \\ 0 &\leq x_{s,z,t}^{\textrm{E,DIS}} \leq U_{s,z,t-1}^{\textrm{E,STO}}*\eta_{s,z}^{\textrm{E,DIS}} \quad \forall s \in \mathcal{S}, z \in \mathcal{Z}, t \in \mathcal{T}^{interior} \\ 0 &\leq x_{s,z,t}^{\textrm{E,DIS}} \leq U_{s,z,t+\tau^{period}-1}^{\textrm{E,STO}}*\eta_{s,z}^{\textrm{E,DIS}} \quad \forall s \in \mathcal{S}, z \in \mathcal{Z}, t \in \mathcal{T}^{start} \end{aligned}\]

DOLPHYN.storage_all_reserves — Method

storage_all_reserves(EP::Model, inputs::Dict)