Storage Asymmetric

DOLPHYN.storage_asymmetric — Method

storage_asymmetric(EP::Model, inputs::Dict, Reserves::Int)

Sets up variables and constraints specific to storage resources with asymmetric charge and discharge capacities.

For storage technologies with asymmetric charge and discharge capacities (all $s \in \mathcal{S}^{asym}$), charge rate $x_{s,z,t}^{\textrm{E,CHA}}$, is constrained by the total installed charge capacity $y_{s,z}^{\textrm{E,STO,CHA}}$, as follows:

\[\begin{equation*} 0 \leq x_{s,z,t}^{\textrm{E,CHA}} \leq y_{s,z}^{\textrm{E,STO,CHA}} \quad \forall s \in \mathcal{S}^{asym}, z \in \mathcal{Z}, t \in \mathcal{T} \end{equation*}\]

If reserves are modeled, the above constraint is replaced by the following:

\[\begin{equation*} 0 \leq x_{s,z,t}^{\textrm{E,CHA}} + f_{s,z,t}^{\textrm{E,CHA}} \leq y_{s,z}^{\textrm{E,STO,CHA}} \quad \forall s \in \mathcal{S}^{asym}, z \in \mathcal{Z}, t \in \mathcal{T} \end{equation*}\]

where $f_{s,z,t}^{\textrm{E,CHA}}$ is the contribution of storage resources to frequency regulation while charging.

DOLPHYN.storage_asymmetric_reserves — Method

storage_asymmetric_reserves(EP::Model, inputs::Dict)

Sets up variables and constraints specific to storage resources with asymmetric charge and discharge capacities when reserves are modeled.

If reserves are modeled, two pairs of proxy variables $f_{s,z,t}^{\textrm{E,CHA}}, f_{s,z,t}^{\textrm{E,DIS}}$ and $r_{s,z,t}^{\textrm{E,CHA}}, r_{s,z,t}^{\textrm{E,DIS}}$ are created for storage resources, to denote the contribution of storage resources to regulation or reserves while charging or discharging, respectively. The total contribution to regulation and reserves, $f_{s,z,t}^{\textrm{E,STO}}, r_{s,z,t}^{\textrm{E,STO}}$ is then the sum of the proxy variables:

\[\begin{aligned} f_{s,z,t}^{\textrm{E,STO}} &= f_{s,z,t}^{\textrm{E,CHA}} + f_{s,z,t}^{\textrm{E,DIS}} \quad \forall s \in \mathcal{S}, z \in \mathcal{Z}, t \in \mathcal{T} \\ r_{s,z,t}^{\textrm{E,STO}} &= r_{s,z,t}^{\textrm{E,CHA}} + r_{s,z,t}^{\textrm{E,DIS}} \quad \forall s \in \mathcal{S}, z \in \mathcal{Z}, t \in \mathcal{T} \end{aligned}\]

The total storage contribution to frequency regulation $f_{s,z,t}^{\textrm{E,STO}}$ and reserves $r_{s,z,t}^{\textrm{E,STO}}$ are each limited specified fraction of installed discharge power capacity $\upsilon^{reg}_{s,z}, \upsilon^{rsv}_{s,z}$), reflecting the maximum ramp rate for the storage resource in whatever time interval defines the requisite response time for the regulation or reserve products (e.g., 5 mins or 15 mins or 30 mins). These response times differ by system operator and reserve product, and so the user should define these parameters in a self-consistent way for whatever system context they are modeling.

\[\begin{aligned} f_{s,z,t}^{\textrm{E,STO}} &\leq \upsilon^{reg}_{s,z} \times y_{s,z}^{\textrm{E,STO,POW}} \forall s \in \mathcal{W}, z \in \mathcal{Z}, t \in \mathcal{T} \\ r_{s,z,t}^{\textrm{E,STO}} &\leq \upsilon^{rsv}_{s,z} \times y_{s,z}^{\textrm{E,STO,POW}} \forall s \in \mathcal{W}, z \in \mathcal{Z}, t \in \mathcal{T} \end{aligned}\]

When charging, reducing the charge rate is contributing to upwards reserve and frequency regulation as it drops net demand. As such, the sum of the charge rate plus contribution to regulation and reserves up must be greater than zero. Additionally, the discharge rate plus the contribution to regulation must be greater than zero.

\[\begin{aligned} 0 \leq x_{s,z,t}^{\textrm{E,CHA}} - f_{s,z,t}^{\textrm{E,CHA}} - r_{s,z,t}^{\textrm{E,CHA}} \quad \forall s \in \mathcal{S}, z \in \mathcal{Z}, t \in \mathcal{T} \\ 0 \leq x_{s,z,t}^{\textrm{E,DIS}} - f_{s,z,t}^{\textrm{E,DIS}} \quad \forall s \in \mathcal{S}, z \in \mathcal{Z}, t \in \mathcal{T} \end{aligned}\]

Additionally, when reserves are modeled, the maximum charge rate and contribution to regulation while charging can be no greater than the available energy storage capacity, or the difference between the total energy storage capacity $y_{s,z}^{\textrm{E,STO},ENE}$, and the state of charge at the end of the previous time period $U_{s,z,t-1}^{\textrm{E,STO}}$. Note that for storage to contribute to reserves down while charging, the storage device must be capable of increasing the charge rate (which increase net load).

\[\begin{equation*} x_{s,z,t}^{\textrm{E,CHA}} + f_{s,z,t}^{\textrm{E,CHA}} \leq y_{s,z}^{\textrm{E,STO,ENE}} - U_{s,z,t-1}^{\textrm{E,STO}} \quad \forall s \in \mathcal{O}, z \in \mathcal{Z}, t \in \mathcal{T} \end{equation*}\]

Finally, the constraints on maximum discharge rate are replaced by the following, to account for capacity contributed to regulation and reserves:

\[\begin{aligned} x_{s,z,t}^{\textrm{E,DIS}} + f_{s,z,t}^{\textrm{E,DIS}} + r_{s,z,t}^{\textrm{E,DIS}} &\leq y_{s,z}^{\textrm{E,STO,POW}} \quad \forall s \in \mathcal{O}, z \in \mathcal{Z}, t \in \mathcal{T} \\ x_{s,z,t}^{\textrm{E,DIS}} + f_{s,z,t}^{\textrm{E,DIS}} + r_{s,z,t}^{\textrm{E,DIS}} &\leq U_{s,z,t-1}^{\textrm{E,STO}} \quad \forall s \in \mathcal{O}, z \in \mathcal{Z}, t \in \mathcal{T} \end{aligned}\]