writeHSCoutputs

DOLPHYN.write_HSC_outputs — Method

write_HSC_outputs(EP::Model, path::AbstractString, setup::Dict, inputs::Dict)

Function for the entry-point for writing the different output files. From here, onward several other functions are called, each for writing specific output files, like costs, capacities, etc.

writeg2pcapacity

writeh2balance

DOLPHYN.write_h2_balance — Method

write_h2_balance(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

writeh2capacity

DOLPHYN.write_h2_capacity — Method

write_h2_capacity(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for writing the diferent capacities for the different generation technologies (starting capacities or, existing capacities, retired capacities, and new-built capacities).

writeh2charge

DOLPHYN.write_h2_charge — Method

write_h2_charge(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for writing the h2 storage charging energy values of the different storage technologies.

writeh2costs

DOLPHYN.write_h2_costs — Method

write_h2_costs(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for writing the costs pertaining to the objective function (fixed, variable O&M etc.).

writeh2emissions

DOLPHYN.write_h2_emissions — Method

write_h2_emissions(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting time-dependent CO$_2$ emissions by zone.

writeh2g2p

DOLPHYN.write_h2_g2p — Method

write_h2_g2p(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for writing the different values of power generated by the different technologies in operation.

writeh2gen

DOLPHYN.write_h2_gen — Method

write_h2_gen(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for writing the different values of power generated by the different technologies in operation.

writeh2nse

DOLPHYN.write_h2_nse — Method

write_h2_nse(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting non-served energy for every model zone, time step and cost-segment.

writeh2pipeline_expansion

DOLPHYN.write_h2_pipeline_expansion — Method

write_h2_pipeline_expansion(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

writeh2pipeline_flow

DOLPHYN.write_h2_pipeline_flow — Method

write_h2_pipeline_flow(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

writeh2storage

DOLPHYN.write_h2_storage — Method

write_h2_storage(path::AbstractString, sep::AbstractString, inputs::Dict,setup::Dict, EP::Model)

Function for writing the capacities of different H2 storage technologies, including hydro reservoir, flexible storage tech etc.

writeh2truck_capacity

DOLPHYN.write_h2_truck_capacity — Method

write_h2_truck_capacity(path::AbstractString, sep::AbstractString, inputs::Dict,setup::Dict, EP::Model)

writeh2truck_flow

DOLPHYN.write_h2_truck_flow — Method

write_h2_truck_flow(path::AbstractString, sep::AbstractString, inputs::Dict,setup::Dict, EP::Model)

writepg2p

DOLPHYN.write_p_g2p — Method

write_p_g2p(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for writing the different values of power generated by the different technologies in operation.