Liquid Fuels Supply Chain Results

Dolphyn.write_liquid_fuels_outputs — Method

writeliquidfuelsoutputs(EP::Model, genxpath::AbstractString, setup::Dict, inputs::Dict)

Function (entry-point) for reporting the different output files of liquid fuels supply chain. From here, onward several other functions are called, each for writing specific output files, like costs, capacities, etc.

Write Liquid Fuels Supply Chain Costs

Dolphyn.write_synfuel_costs — Method

write_synfuel_costs(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for writing the cost for the different sectors of the liquid fuels supply chain (Synthetic fuel resources CAPEX and OPEX, different types of conventional gasoline, jetfuel, and diesel).

Write Synthetic Fuels Resources Capacity

Dolphyn.write_synfuel_capacity — Method

write_synfuel_capacity(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for writing the diferent capacities for synthetic fuels resources.

Write Synthetic Fuels Resources Process Balance

Dolphyn.write_synfuel_balance — Method

write_synfuel_balance(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting input and output balance of synthetic fuels resources across different zones with time.

Write Synthetic Fuels Production Balance

Dolphyn.write_synfuel_gen — Method

write_synfuel_gen(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting the different values of synthetic gasoline, jetfuel, and diesel generated by the different technologies in operation.

Write Liquid Fuels Emissions

Dolphyn.write_synfuel_emissions — Method

write_synfuel_emissions(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting CO2 emissions of different liquid fuel types across different zones.

Write Liquid Fuels Demand Balance

Dolphyn.write_liquid_fuel_demand_balance — Method

write_liquid_fuel_demand_balance(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting the liquid fuels balance of resources across different zones with time for each type of fuels.