h2g2pall

DOLPHYN.h2_g2p_all — Method

h2_g2p_all(EP::Model, inputs::Dict, setup::Dict)

The h2 generation module creates decision variables, expressions, and constraints related to hydrogen generation infrastructure

Investment and FOM cost expression, VOM cost expression, minimum and maximum capacity limits

h2g2pcommit

Missing docstring.

Missing docstring for DOLPHYN.h2_g2p_commit(EP::Model, inputs::Dict, setup::Dict). Check Documenter's build log for details.

h2g2pdischarge

Missing docstring.

Missing docstring for DOLPHYN.h2_g2p_discharge(EP::Model, inputs::Dict, setup::Dict). Check Documenter's build log for details.

h2g2pinvestment

DOLPHYN.h2_g2p_investment — Method

h2_discharge(EP::Model, inputs::Dict, UCommit::Int, Reserves::Int)

This module defines the production decision variable representing hydrogen injected into the network by resource $y$ by at time period $t$.

This module additionally defines contributions to the objective function from variable costs of generation (variable O&M plus fuel cost) from all resources over all time periods.

h2g2pno_commit

Missing docstring.

Missing docstring for DOLPHYN.h2_g2p_no_commit(EP::Model, inputs::Dict,setup::Dict). Check Documenter's build log for details.

h2_g2p

DOLPHYN.h2_g2p — Method

H2_G2Peration(EP::Model, inputs::Dict, UCommit::Int, Reserves::Int)

The h2_production module creates decision variables, expressions, and constraints related to various hydrogen generation technologies (electrolyzers, natural gas reforming etc.)

This module uses the following 'helper' functions in separate files: H2_G2Peration_commit() for resources subject to unit commitment decisions and constraints (if any) and H2_G2Peration_no_commit() for resources not subject to unit commitment (if any).