Hydrogen Gas to Power

Hydrogen Gas to Power All

DOLPHYN.h2_g2p_all — Method

h2_g2p_all(EP::Model, inputs::Dict, setup::Dict)

The hydrogen to power module creates decision variables, expressions, and constraints related to hydrogen generation infrastructure

Investment and FOM cost expression, VOM cost expression, minimum and maximum capacity limits

One cannot retire more capacity than existing capacity.

\[\begin{aligned} &\Delta_{y,z} \leq \overline{\Delta_{y,z}} \hspace{4 cm} \forall y \in \mathcal{G}, z \in \mathcal{Z} \end{aligned}\]

For resources where $\overline{\Omega_{y,z}}$ and $\underline{\Omega_{y,z}}$ is defined, then we impose constraints on minimum and maximum power capacity.

\[\begin{aligned} & \Delta^{total}_{y,z} \leq \overline{\Omega}_{y,z} \hspace{4 cm} \forall y \in \mathcal{G}, z \in \mathcal{Z} \\ & \Delta^{total}_{y,z} \geq \underline{\Omega}_{y,z} \hspace{4 cm} \forall y \in \mathcal{G}, z \in \mathcal{Z} \end{aligned}\]

Hydrogen Gas to Power Commitment

DOLPHYN.h2_g2p_commit — Method

h2_g2p_commit(EP::Model, inputs::Dict, setup::Dict)

This module creates decision variables, expressions, and constraints related to various hydrogen to power technologies with unit commitment constraints (e.g. natural gas reforming etc.)

G2P Unit commitment decision variables:

This function defines the following decision variables:

\[\nu_{y,t,z}\]

designates the commitment state of g2p generator cluster $h$ in zone $z$ at time $t$; $\chi_{y,t,z}$ represents number of g2p startup decisions in cluster $h$ in zone $z$ at time $t$; $\zeta_{y,t,z}$ represents number of g2p shutdown decisions in cluster $h$ in zone $z$ at time $t$.

Cost expressions:

The total cost of start-ups across g2p generators subject to unit commitment ($h \in UC$) and all time periods, t is expressed as:

\[\begin{aligned} C^{start} = \sum_{h \in UC, t \in T} \omega_t \times start\_cost_{h} \times \chi_{h,t} \end{aligned}\]

The sum of start-up costs is added to the objective function.

Startup and shutdown events (thermal plant cycling)

Capacitated limits on g2p unit commitment decision variables

Hydrogen to power resources subject to unit commitment ($h \in \mathcal{UC}$) adhere to the following constraints on commitment states, startup events, and shutdown events, which limit each decision to be no greater than the maximum number of discrete units installed (as per the following three constraints):

\[\begin{aligned} \nu_{h,z,t} \leq \frac{\Delta^{\text{total}}_{h,z}}{\Omega^{size}_{h,z}} \hspace{1.5cm} \forall h \in \mathcal{UC}, \forall z \in \mathcal{Z}, \forall t \in \mathcal{T} \end{aligned}\]

\[\begin{aligned} \chi_{h,z,t} \leq \frac{\Delta^{\text{total}}_{h,z}}{\Omega^{size}_{h,z}} \hspace{1.5cm} \forall h \in \mathcal{UC}, \forall z \in \mathcal{Z}, \forall t \in \mathcal{T} \end{aligned}\]

\[\begin{aligned} \zeta_{h,z,t} \leq \frac{\Delta^{\text{total}}_{h,z}}{\Omega^{size}_{h,z}} \hspace{1.5cm} \forall h \in \mathcal{UC}, \forall z \in \mathcal{Z}, \forall t \in \mathcal{T} \end{aligned}\]

(See Constraints 1-3 in the code)

where decision $\nu_{h,z,t}$ designates the g2p commitment state of generator cluster $h$ in zone $z$ at time $t$, decision $\chi_{h,z,t}$ represents number of startup decisions, decision $\zeta_{h,z,t}$ represents number of shutdown decisions, $\Delta^{\text{total}}_{h,z}$ is the total installed capacity, and parameter $\Omega^{size}_{h,z}$ is the unit size.

Hydrogen to power commitment state constraint linking start-up and shut-down decisions

Additionally, the following constarint maintains the commitment state variable across time, $\nu_{h,z,t}$, as the sum of the commitment state in the prior, $\nu_{h,z,t-1}$, period plus the number of units started in the current period, $\chi_{h,z,t}$, less the number of units shut down in the current period, $\zeta_{h,z,t}$:

\[\begin{aligned} &\nu_{h,z,t} =\nu_{h,z,t-1} + \chi_{h,z,t} - \zeta_{h,z,t} \hspace{1.5cm} \forall h \in \mathcal{UC}, \forall z \in \mathcal{Z}, \forall t \in \mathcal{T}^{interior} \\ &\nu_{h,z,t} =\nu_{h,z,t +\tau^{period}-1} + \chi_{h,z,t} - \zeta_{h,z,t} \hspace{1.5cm} \forall h \in \mathcal{UC}, \forall z \in \mathcal{Z}, \forall t \in \mathcal{T}^{start} \end{aligned}\]

(See Constraint 4 in the code)

Like other time-coupling constraints, this constraint wraps around to link the commitment state in the first time step of the year (or each representative period), $t \in \mathcal{T}^{start}$, to the last time step of the year (or each representative period), $t+\tau^{period}-1$.

Ramping constraints

Thermal resources subject to unit commitment ($h \in UC$) adhere to the following ramping constraints on hourly changes in power output:

\[\begin{aligned} \Theta_{h,z,t-1} - \Theta_{h,z,t} &\leq \kappa^{down}_{h,z} \cdot \Omega^{size}_{h,z} \cdot (\nu_{h,z,t} - \chi_{h,z,t}) & \\[6pt] \qquad & - \: \rho^{min}_{h,z} \cdot \Omega^{size}_{h,z} \cdot \chi_{h,z,t} & \hspace{0.5cm} \forall h \in \mathcal{UC}, \forall z \in \mathcal{Z}, \forall t \in \mathcal{T} \\[6pt] \qquad & + \: \text{min}( \rho^{max}_{h,z,t}, \text{max}( \rho^{min}_{h,z}, \kappa^{down}_{h,z} ) ) \cdot \Omega^{size}_{h,z} \cdot \zeta_{h,z,t} & \end{aligned}\]

\[\begin{aligned} \Theta_{h,z,t} - \Theta_{h,z,t-1} &\leq \kappa^{up}_{h,z} \cdot \Omega^{size}_{h,z} \cdot (\nu_{h,z,t} - \chi_{h,z,t}) & \\[6pt] \qquad & + \: \text{min}( \rho^{max}_{h,z,t}, \text{max}(\rho^{min}_{h,z}, \kappa^{up}_{h,z})) \cdot \Omega^{size}_{h,z} \cdot \chi_{h,z,t} & \hspace{0.5cm} \forall h \in \mathcal{UC}, \forall z \in \mathcal{Z}, \forall t \in \mathcal{T} \\[6pt] \qquad & - \: \rho^{min}_{h,z} \cdot \Omega^{size}_{h,z} \cdot \zeta_{h,z,t} & \end{aligned}\]

(See Constraints 5-6 in the code)

where decision $\Theta_{h,z,t}$ is the energy injected into the grid by technology $h$ in zone $z$ at time $t$, parameter $\kappa_{h,z,t}^{up|down}$ is the maximum ramp-up or ramp-down rate as a percentage of installed capacity, parameter $\rho_{h,z}^{min}$ is the minimum stable power output per unit of installed capacity, and parameter $\rho_{h,z,t}^{max}$ is the maximum available generation per unit of installed capacity. These constraints account for the ramping limits for committed (online) units as well as faster changes in power enabled by units starting or shutting down in the current time step.

Minimum and maximum power output

If not modeling regulation and spinning reserves, thermal resources subject to unit commitment adhere to the following constraints that ensure power output does not exceed minimum and maximum feasible levels:

\[\begin{aligned} \Theta_{h,z,t} \geq \rho^{min}_{h,z} \times \Omega^{size}_{h,z} \times \nu_{h,z,t} \hspace{1.5cm} \forall h \in \mathcal{UC}, \forall z \in \mathcal{Z}, \forall t \in \mathcal{T} \end{aligned}\]

\[\begin{aligned} \Theta_{h,z,t} \geq \rho^{max}_{h,z} \times \Omega^{size}_{h,z} \times \nu_{h,z,t} \hspace{1.5cm} \forall h \in \mathcal{UC}, \forall z \in \mathcal{Z}, \forall t \in \mathcal{T} \end{aligned}\]

(See Constraints 7-8 the code)

Minimum and maximum up and down time

Thermal resources subject to unit commitment adhere to the following constraints on the minimum time steps after start-up before a unit can shutdown again (minimum up time) and the minimum time steps after shut-down before a unit can start-up again (minimum down time):

\[\begin{aligned} \nu_{h,z,t} \geq \displaystyle \sum_{\hat{t} = t-\tau^{up}_{h,z}}^t \chi_{h,z,\hat{t}} \hspace{1.5cm} \forall h \in \mathcal{UC}, \forall z \in \mathcal{Z}, \forall t \in \mathcal{T} \end{aligned}\]

\[\begin{aligned} \frac{\overline{\Delta_{h,z}} + \Omega_{h,z} - \Delta_{h,z}}{\Omega^{size}_{h,z}} - \nu_{h,z,t} \geq \displaystyle \sum_{\hat{t} = t-\tau^{down}_{h,z}}^t \zeta_{h,z,\hat{t}} \hspace{1.5cm} \forall h \in \mathcal{UC}, \forall z \in \mathcal{Z}, \forall t \in \mathcal{T} \end{aligned}\]

(See Constraints 9-10 in the code)

where $\tau_{h,z}^{up|down}$ is the minimum up or down time for units in generating cluster $h$ in zone $z$.

Like with the ramping constraints, the minimum up and down constraint time also wrap around from the start of each time period to the end of each period. It is recommended that users of GenX must use longer subperiods than the longest min up/down time if modeling UC. Otherwise, the model will report error.

Hydrogen Gas to Power Discharge

DOLPHYN.h2_g2p_discharge — Method

h2_g2p_discharge(EP::Model, inputs::Dict, setup::Dict)

This module defines the production decision variable representing power form hydrogen injected into the network by resource $y$ by at time period $t$.

This module additionally defines contributions to the objective function from variable costs of generation (variable O&M plus fuel cost) from all resources over all time periods.

\[\begin{aligned} Obj_{Var\_g2p} = \sum_{h \in \mathcal{H}} \sum_{t \in \mathcal{T}}\omega_{t}\times(\pi^{VOM}_{h} \end{aligned}\]

Hydrogen Gas to Power Investment

DOLPHYN.h2_g2p_investment — Method

h2_g2p_investment(EP::Model, inputs::Dict, setup::Dict)

This function defines the expressions and constraints keeping track of total available gas to power capacity.

The total capacity of g2p is defined as the sum of the existing capacity plus the newly invested capacity minus any retired capacity.

\[\begin{aligned} & \Delta^{total}_{h,z} =(\overline{\Delta_{h,z}}+\Omega_{h,z}-\Delta_{h,z}) \forall h \in \mathcal{H}, z \in \mathcal{Z} \end{aligned}\]

In addition, this function adds investment and fixed O\&M related costs related to discharge/generation capacity to the objective function:

\[\begin{aligned} & \sum_{h \in \mathcal{H} } \sum_{z \in \mathcal{Z}} \left( (\pi^{INVEST}_{h,z} \times \overline{\Omega}^{size}_{h,z} \times \Omega_{h,z}) + (\pi^{FOM}_{h,z} \times \overline{\Omega}^{size}_{h,z} \times \Delta^{total}_{h,z})\right) \end{aligned}\]

Hydrogen Gas to Power No Commitment

DOLPHYN.h2_g2p_no_commit — Method

h2_g2p_no_commit(EP::Model, inputs::Dict,setup::Dict)

This module creates decision variables, expressions, and constraints related to various hydrogen to power technologies without unit commitment constraints

Ramping limits

Hydrogen to power resources not subject to unit commitment ($h \in H \setminus UC$) adhere instead to the following ramping limits on hourly changes in power output:

\[\begin{aligned} \Theta_{h,z,t-1} - \Theta_{h,z,t} \leq \kappa_{h,z}^{down} \Delta^{\text{total}}_{h,z} \hspace{1cm} \forall h \in \mathcal{UC}, \forall z \in \mathcal{Z}, \forall t \in \mathcal{T} \end{aligned}\]

\[\begin{aligned} \Theta_{h,z,t} - \Theta_{h,z,t-1} \leq \kappa_{h,z}^{up} \Delta^{\text{total}}_{h,z} \hspace{1cm} \forall h \in \mathcal{UC}, \forall z \in \mathcal{Z}, \forall t \in \mathcal{T} \end{aligned}\]

(See Constraints 1-2 in the code)

This set of time-coupling constraints wrap around to ensure the power output in the first time step of each year (or each representative period), $t \in \mathcal{T}^{start}$, is within the eligible ramp of the power output in the final time step of the year (or each representative period), $t+\tau^{period}-1$.

Minimum and maximum power output

When not modeling regulation and reserves, thermal units not subject to unit commitment decisions are bound by the following limits on maximum and minimum power output:

\[\begin{aligned} \Theta_{h,z,t} \geq \rho^{min}_{h,z} \times \Delta^{total}_{h,z} \hspace{1cm} \forall h \in \mathcal{UC}, \forall z \in \mathcal{Z}, \forall t \in \mathcal{T} \end{aligned}\]

\[\begin{aligned} \Theta_{h,z,t} \leq \rho^{max}_{h,z,t} \times \Delta^{total}_{h,z} \hspace{1cm} \forall h \in \mathcal{UC}, \forall z \in \mathcal{Z}, \forall t \in \mathcal{T} \end{aligned}\]

(See Constraints 3-4 in the code)

Hydrogen Gas to Power

DOLPHYN.h2_g2p — Method

h2_g2p(EP::Model, inputs::Dict, setup::Dict)

This module creates decision variables, expressions, and constraints related to various hydrogen to power technologies as well as carbon emission policy constraints.