Hydrogen Production No Commitment

Dolphyn.h2_production_no_commit — Method

h2_production_no_commit(EP::Model, inputs::Dict,setup::Dict)

This function defines the operating constraints for hydrogen generation plants (thermal and electrolysis) NOT subject to unit commitment constraints on power plant start-ups and shut-down decisions $g \in \mathcal{THE} \setminus \mathcal{UC}$.

Hydrogen balance expressions

Contributions to the hydrogen balance expression from each thermal resources without unit commitment $g \in \mathcal{THE} \setminus \mathcal{UC}$ are also defined below. If liquid hydrogen is modeled, a liquid hydrogen balance expression is needed and contributions to the gas balance are accounted for.

\[\begin{equation*} HydrogenBalGas_{GEN} = \sum_{g \in \mathcal{G}} x_{g,z,t}^{\textrm{H,GEN}} - \sum_{g \in \mathcal{G}} x_{g,z,t}^{\textrm{H,LIQ}} + \sum_{g \in \mathcal{G}} x_{g,z,t}^{\textrm{H,EVAP}} \forall z \in \mathcal{Z}, t \in \mathcal{T} \end{equation*}\]

\[\begin{equation*} HydrogenBalLiq_{GEN} = \sum_{g \in \mathcal{G}} x_{g,z,t}^{\textrm{H,LIQ}} - \sum_{g \in \mathcal{G}} x_{g,z,t}^{\textrm{H,EVAP}} \forall z \in \mathcal{Z}, t \in \mathcal{T} \end{equation*}\]

Ramping limits

Thermal resources not subject to unit commitment $k \in \mathcal{THE} \setminus \mathcal{UC}$ adhere instead to the following ramping limits on hourly changes in hydrogen output:

\[\begin{equation*} x_{g,z,t-1}^{\textrm{H,GEN}} - x_{g,z,t}^{\textrm{H,GEN}} \leq \kappa_{g,z}^{\textrm{H,DN}} y_{g,z}^{\textrm{H,GEN}} \quad \forall g \in \mathcal{THE} \setminus \mathcal{UC}, z \in \mathcal{Z}, t \in \mathcal{T} \end{equation*}\]

\[\begin{equation*} x_{g,z,t}^{\textrm{H,GEN}} - x_{g,z,t-1}^{\textrm{H,GEN}} \leq \kappa_{g,z}^{\textrm{H,UP}} y_{g,z}^{\textrm{H,GEN}} \quad \forall g \in \mathcal{THE} \setminus \mathcal{UC}, z \in \mathcal{Z}, t \in \mathcal{T} \end{equation*}\]

(See Constraints 1-2 in the code)

This set of time-coupling constraints wrap around to ensure the hydrogen output in the first time step of each year (or each representative period), $t \in \mathcal{T}^{start}$, is within the eligible ramp of the power output in the final time step of the year (or each representative period), $t+\tau^{period}-1$.

Minimum and maximum hydrogen output

\[\begin{equation*} x_{g,z,t}^{\textrm{H,GEN}} \geq \underline{R_{g,z}^{\textrm{H,GEN}}} \times y_{g,z}^{\textrm{H,GEN}} \quad \forall g \in \mathcal{THE} \setminus \mathcal{UC}, z \in \mathcal{Z}, t \in \mathcal{T} \end{equation*}\]

\[\begin{equation*} x_{g,z,t}^{\textrm{H,GEN}} \leq \overline{R_{g,z}^{\textrm{H,GEN}}} \times y_{g,z}^{\textrm{H,GEN}} \quad \forall g \in \mathcal{THE} \setminus \mathcal{UC}, z \in \mathcal{Z}, t \in \mathcal{T} \end{equation*}\]

(See Constraints 3-4 in the code)