Load Hydrogen Inputs

Dolphyn.load_h2_inputs — Method

load_h2_inputs(inputs::Dict,setup::Dict,path::AbstractString)

Loads various data inputs from multiple input .csv files in path directory and stores variables in a Dict (dictionary) object for use in model() function

inputs: inputs - dict object containing input data setup - dict object containing setup parameters path - string path to working directory

returns: Dict (dictionary) object containing all data inputs of hydrogen sector.

Load Carbon Cap in Hydrogen Sector

Dolphyn.load_co2_cap_hsc — Method

load_co2_cap_hsc(setup::Dict, path::AbstractString, sep::AbstractString, inputs_co2_hsc::Dict)

Function for reading input parameters related to CO$_2$ emissions cap constraints in hydrogen supply chain only.

Load Hydrogen Demand

Dolphyn.load_h2_demand — Method

load_h2_demand(setup::Dict, path::AbstractString, sep::AbstractString, inputs_load::Dict)

Function for reading input parameters related to hydrogen load (demand) of each zone.

Load Hydrogen Gas to Power Variability

Dolphyn.load_h2_g2p_variability — Method

load_h2_g2p_variability(setup::Dict, path::AbstractString, sep::AbstractString, inputs_genvar::Dict)

Function for reading input parameters related to hourly maximum capacity factors for all hydrogen to power plants.

Load Hydrogen Gas to Power

Dolphyn.load_h2_g2p — Method

load_h2_g2p(setup::Dict, path::AbstractString, sep::AbstractString, inputs_gen::Dict)

Function for reading input parameters related to hydrogen to power generators.

Load Hydrogen Generators

Dolphyn.load_h2_gen — Method

load_h2_gen(setup::Dict, path::AbstractString, sep::AbstractString, inputs_gen::Dict)

Function for reading input parameters related to hydrogen generators. Liquifiers and evaporators are also considered here, and are identified via the H2LIQ type column. Liquid versus gas storage is identified via the H2STOR type column.

Load Hydrogen Generators Variability

Dolphyn.load_h2_generators_variability — Method

load_h2_generators_variability(setup::Dict, path::AbstractString, sep::AbstractString, inputs_genvar::Dict)

Function for reading input parameters related to hourly maximum capacity factors for all hydrogen generators (plus storage).

Load Hydrogen Pipelines

Dolphyn.load_h2_pipeline_data — Method

load_h2_pipeline_data(setup::Dict, path::AbstractString, sep::AbstractString, inputs_nw::Dict)

Function for reading input parameters related to the hydrogen transmission network via pipelines.

Load Hydrogen Truck

Dolphyn.load_h2_truck — Method

load_h2_truck(path::AbstractString, sep::AbstractString, inputs_truck::Dict)

Function for reading input parameters related to hydrogen trucks. A variable is created to distinguish between Gas and Liquid trucks, which is relevant for the hydrogen balance expressions. Other truck types like LOHC are currently not considered, but may need to be identified as Gas for the balance.