Hydrogen Supply Chain Results

DOLPHYN.write_HSC_outputs — Method

write_HSC_outputs(EP::Model, path::AbstractString, setup::Dict, inputs::Dict)

Function (entry-point) for reporting the different output files of hydrogen supply chain. From here, onward several other functions are called, each for writing specific output files, like costs, capacities, etc.

Write Hydrogen Gas to Power Capacity

DOLPHYN.write_g2p_capacity — Method

write_g2p_capacity(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting the diferent capacities for the different hydrogen to power technologies (starting capacities or, existing capacities, retired capacities, and new-built capacities).

Write Hydrogen Gas to Power Balance

DOLPHYN.write_h2_g2p — Method

write_h2_g2p(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting the different values of power generated by hydrogen to power plants in operation.

Write Hydrogen Balance

DOLPHYN.write_h2_balance — Method

write_h2_balance(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting hydrogen balance.

Write Hydrogen Capacity

DOLPHYN.write_h2_capacity — Method

write_h2_capacity(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting the capacities for the different hydrogen resources (starting capacities or, existing capacities, retired capacities, and new-built capacities).

Write Hydrogen Charge

DOLPHYN.write_h2_charge — Method

write_h2_charge(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting the h2 storage charging values of the different hydrogen storage technologies.

Write Hydrogen Costs

DOLPHYN.write_h2_costs — Method

write_h2_costs(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting the costs of hydrogen supply chain pertaining to the objective function (fixed, variable O&M etc.).

Write Hydrogen Emissions

DOLPHYN.write_h2_emissions — Method

write_h2_emissions(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting time-dependent CO$_2$ emissions by zone in hydrogen supply chain.

Write Hydrogen Generation

DOLPHYN.write_h2_gen — Method

write_h2_gen(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting the different values of hydrogen generated by the different technologies in operation.

Write Non-served Hydrogen

DOLPHYN.write_h2_nse — Method

write_h2_nse(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting non-served hydrogen for every model zone, time step and cost-segment.

Write Hydrogen Pipeline Expansion

DOLPHYN.write_h2_pipeline_expansion — Method

write_h2_pipeline_expansion(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting the expansion of hydrogen pipelines.

Write Hydrogen Pipeline Flow

DOLPHYN.write_h2_pipeline_flow — Method

write_h2_pipeline_flow(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting the hydrogen flow via pipeliens.

Write Hydrogen Storage

DOLPHYN.write_h2_storage — Method

write_h2_storage(path::AbstractString, sep::AbstractString, inputs::Dict,setup::Dict, EP::Model)

Function for reporting the capacities of different hydrogen storage technologies, including hydro reservoir, flexible storage tech etc.

Write Hydrogen Truck Capacity

DOLPHYN.write_h2_truck_capacity — Method

write_h2_truck_capacity(path::AbstractString, sep::AbstractString, inputs::Dict,setup::Dict, EP::Model)

Functions for reporting capacities of hydrogen trucks (starting capacities or, existing capacities, retired capacities, and new-built capacities).

Write Hydrogen Truck Flow

DOLPHYN.write_h2_truck_flow — Method

write_h2_truck_flow(path::AbstractString, sep::AbstractString, inputs::Dict,setup::Dict, EP::Model)

Fucntion for reporting hydrogen flow via trucsk.

Write hydrogen Gas to Power

DOLPHYN.write_p_g2p — Method

write_p_g2p(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting the different values of power generated by hydrogen to power plants.