Functions for Writing the Different Results/Outputs to Separate Files

DOLPHYN.write_outputs — Method

write_outputs(EP::Model, path::AbstractString, setup::Dict, inputs::Dict)

Function (entry-point) for reporting the different output files. From here, onward several other functions are called, each for writing specific output files, like costs, capacities, etc.

Write Status

DOLPHYN.write_status — Method

write_status(path::AbstractString, sep::AbstractString, inputs::Dict, EP::Model)

Function for reporting the final solve status of the optimization problem solved.

Write Costs

DOLPHYN.write_costs — Method

write_costs(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting the costs pertaining to the objective function (fixed, variable O&M etc.).

Write Emissions

DOLPHYN.write_emissions — Method

write_emissions(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting time-dependent CO$_2$ emissions by zone.

Write Capacities

DOLPHYN.write_capacity — Method

write_capacity(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model))

Function for reporting the diferent capacities for the different generation technologies (starting capacities or, existing capacities, retired capacities, and new-built capacities).

Write Capacity Values

DOLPHYN.write_capacity_value — Method

write_capacity_value(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, dfPower::DataFrame, dfCharge::DataFrame, dfResMar::DataFrame, dfCap::DataFrame)

Function for reporting the capacity value of generation and storage resources to csv file.

Write Capacity Reserve

DOLPHYN.write_reserve_margin — Method

write_reserve_margin(path::AbstractString, sep::AbstractString, setup::Dict, EP::Model)

Function for reporting capacity reserve margin dual.

DOLPHYN.write_reserve_margin_w — Method

write_reserve_margin_w(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting capacity reserve margin dual modified by weights.

Write Charge Values

DOLPHYN.write_charge — Method

write_charge(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting the charging energy values of the different storage technologies.

Write Charge Costs

DOLPHYN.write_charging_cost — Method

write_charging_cost(path::AbstractString, sep::AbstractString, inputs::Dict, dfCharge::DataFrame, dfPrice::DataFrame, dfPower::DataFrame, setup::Dict)

Function for reporting charging costs of each resource.

Write Non-Served Energy

DOLPHYN.write_nse — Method

write_nse(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting non-served energy for every model zone, time step and cost-segment.

Write Storage

DOLPHYN.write_storage — Method

write_storage(path::AbstractString, sep::AbstractString, inputs::Dict,setup::Dict, EP::Model)

Function for reporting the capacities of different storage technologies, including hydro reservoir, flexible storage tech etc.

Write Storage Dual

DOLPHYN.write_storagedual — Method

write_storagedual(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting dual of storage level (state of charge) balance of each resource in each time step.

Write Long Duration Storage

DOLPHYN.write_opwrap_lds_dstor — Method

write_opwrap_lds_dstor(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting long duration storage related operations.

Write Power

DOLPHYN.write_power — Method

write_power(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting the different values of power generated by the different technologies in operation.

Write Power Balance

DOLPHYN.write_power_balance — Method

write_power_balance(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting power balance of resources across different zones.

Write Unit Commitment

DOLPHYN.write_commit — Method

write_commit(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting committment states for each resource at each time step.

DOLPHYN.write_start — Method

write_start(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Fucntion for reporting startup action of resources at each time step.

DOLPHYN.write_shutdown — Method

write_shutdown(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting the shutdown action of resources at each time step.

Write Curtailment

DOLPHYN.write_curtailment — Method

write_curtailment(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting the curtailment values of the different variable renewable resources.

Write Network Expansion

DOLPHYN.write_nw_expansion — Method

write_nw_expansion(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting the expansion statuses of electricity network expansion.

Write Network Flow

DOLPHYN.write_transmission_flows — Method

write_transmission_flows(path::AbstractString, sep::AbstractString, setup::Dict, inputs::Dict, EP::Model)

Function for reporting electricity flow via transmision lines.

Write Network Losses

DOLPHYN.write_transmission_losses — Method

write_transmission_losses(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting electricity losses during transmission via lines.

Write Prices

DOLPHYN.write_price — Method

write_price(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting marginal electricity price for each model zone and time step. Marginal electricity price is equal to the dual variable of the load balance constraint. If GenX is configured as a mixed integer linear program, then this output is only generated if WriteShadowPrices flag is activated. If configured as a linear program (i.e. linearized unit commitment or economic dispatch) then output automatically available.

Write Reliability

DOLPHYN.write_reliability — Method

write_reliability(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting dual variable of maximum non-served energy constraint (shadow price of reliability constraint) for each model zone and time step.

Write Reserves

DOLPHYN.write_reg — Method

write_reg(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting the regulation action.

DOLPHYN.write_rsv — Method

write_rsv(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting the reserve action operations.

Write Energy Revenue

DOLPHYN.write_energy_revenue — Method

write_energy_revenue(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model, dfPower::DataFrame, dfPrice::DataFrame, dfCharge::DataFrame)

Function for reporting energy revenue from the different generation technologies.

Write Subsidy Revenue

DOLPHYN.write_subsidy_revenue — Method

write_subsidy_revenue(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, dfCap::DataFrame, EP::Model)

Function for reporting subsidy revenue earned if a generator specified Min_Cap is provided in the input file. GenX will print this file only the shadow price can be obtained form the solver. Do not confuse this with the Minimum Capacity Carveout constraint, which is for a subset of generators, and a separate revenue term will be calculated in other files. The unit is $.

Write Capacity Revenue

DOLPHYN.write_reserve_margin_revenue — Method

write_reserve_margin_revenue(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, dfPower::DataFrame, dfCharge::DataFrame, dfResMar::DataFrame, dfCap::DataFrame)

Function for reporting the capacity revenue earned by each generator listed in the input file. GenX will print this file only when capacity reserve margin is modeled and the shadow price can be obtained form the solver. Each row corresponds to a generator, and each column starting from the 6th to the second last is the total revenue from each capacity reserve margin constraint. The revenue is calculated as the capacity contribution of each time steps multiplied by the shadow price, and then the sum is taken over all modeled time steps. The last column is the total revenue received from all capacity reserve margin constraints. As a reminder, GenX models the capacity reserve margin (aka capacity market) at the time-dependent level, and each constraint either stands for an overall market or a locality constraint.

Write Energy Share Requirement Revenue

DOLPHYN.write_esr_revenue — Method

write_esr_revenue(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, dfPower::DataFrame, dfESR::DataFrame)

Function for reporting the renewable/clean credit revenue earned by each generator listed in the input file. GenX will print this file only when RPS/CES is modeled and the shadow price can be obtained form the solver. Each row corresponds to a generator, and each column starting from the 6th to the second last is the total revenue earned from each RPS constraint. The revenue is calculated as the total annual generation (if elgible for the corresponding constraint) multiplied by the RPS/CES price. The last column is the total revenue received from all constraint. The unit is $.

Write Energy Share Requirement Prices

DOLPHYN.write_esr_prices — Method

write_esr_prices(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for reporting prices under energy share requirements.

Write Net Revenue

DOLPHYN.write_net_revenue — Method

write_net_revenue(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model, dfCap::DataFrame, dfESRRev::DataFrame, dfResRevenue::DataFrame, dfChargingcost::DataFrame, dfPower::DataFrame, dfEnergyRevenue::DataFrame, dfSubRevenue::DataFrame, dfRegSubRevenue::DataFrame)

Function for reporting network revenue of different generation technologies.

Write Time Weights

DOLPHYN.write_time_weights — Method

write_time_weights(path::AbstractString, sep::AbstractString, inputs::Dict)

Function for reporting the time weights after clustering process.